% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.2D.gaussian.density_function.R
\name{calc.2D.gaussian.density}
\alias{calc.2D.gaussian.density}
\title{Calculate gaussian density for a point in 2D space}
\usage{
calc.2D.gaussian.density(w, Xctr, Yctr, Xcor, Ycor)
}
\arguments{
\item{w}{width of the Gauss distribution}

\item{Xctr}{X-coordinate of the center point}

\item{Yctr}{Y-coordinate of the center point}

\item{Xcor}{X-coordinate of the point}

\item{Ycor}{Y-coordinate of the point}
}
\value{
Vector of density values for each input point
}
\description{
Function that returns for each X-Y-coordinate pair, the gaussian density around
a given center point and with a given width.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{2D}
\keyword{Gaussian}
\keyword{density}
\keyword{distribution}
\keyword{weighting}

