% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.cumulative.size.index_function.R
\name{calc.cumulative.size.index}
\alias{calc.cumulative.size.index}
\title{A function to calculate the CSI cummulative Size Index}
\usage{
calc.cumulative.size.index(dbh, h_max, Xfactor = 0.6, part2 = "old")
}
\arguments{
\item{dbh, }{vector of all dbh of a plot in m}

\item{h_max, }{height of the talles tree in the plot}

\item{Xfactor, }{which Fraction of the cummulative dbh should be devided by the cummulative dbh}

\item{part2, }{"no" returns only part 1, "dbh_max" multilpies part1 with dbhmax^2.5; "h_max" multiplies part 1 with f(h_max,dbh_max);"old" with h_max^1.5}
}
\value{
dimensionless CSI
}
\description{
A function to calculate the CSI cummulative Size Index
}
\examples{
in progress
}
\author{
Friedrich J. Bohn
}
\keyword{cumulative}
\keyword{index}
\keyword{size}

