% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.discrete.Pareto.lambda_function.R
\name{calc.discrete.Pareto.lambda}
\alias{calc.discrete.Pareto.lambda}
\title{Calculate scaling parameter lambda of a discrete power law distribution}
\usage{
calc.discrete.Pareto.lambda(vec)
}
\arguments{
\item{vec}{vector of (discrete) size measurements}
}
\value{
vector of two elements: [1] lambda, [2] two-times the negative log-likelihood of the ML-fit
}
\description{
Function that returns the scaling parameter lambda of a
discrete Pareto (power law) distribution (e.g. canopy gap sizes,
tree diameters) using Maximum-Likelihood estimation (MLE);
Code adapted from Asner et al. (2013) supplements.
}
\examples{
in progress
}
\keyword{Pareto}
\keyword{discrete}
\keyword{distribution}
\keyword{lambda}
\keyword{law}
\keyword{power}
\keyword{size}

