% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.FCC.from.point.cloud_function.R
\name{calc.FCC.from.point.cloud}
\alias{calc.FCC.from.point.cloud}
\title{Calculate fractional canopy cover (FCC) from XYZ-dataframe}
\usage{
calc.FCC.from.point.cloud(XYZdf, pxl.res = 1, h.threshold = 1)
}
\arguments{
\item{XYZdf}{Point cloud in XYZ-dataframe format}

\item{pxl.res}{Side length of CHM pixel}

\item{h.threshold}{Height threshold (pixels with the exact threshold value are included as canopy)}
}
\value{
Fractional canopy cover (FCC) as a value between 0 and 1
}
\description{
Function to calculate FCC from the point cloud in XYZ-dataframe format for a
custom pixel size and height threshold.
FCC is the relative share of all CHM pixels values above a height threshold.
The CHM is generated internally from the point cloud by rasterizing the highest
return in each pixel of a given resolution.
}
\examples{
in progress
}
\keyword{CHM}
\keyword{FCC}
\keyword{Lidar}
\keyword{XYZ}
\keyword{canopy}
\keyword{cover}
\keyword{forest}
\keyword{fraction}
\keyword{fractional}
\keyword{gap}
\keyword{height}
\keyword{pixel}
\keyword{proportion}
\keyword{raster}
\keyword{relative}
\keyword{resolution}
\keyword{threshold}
