% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.field.horiz.index.TUM_function.R
\name{calc.field.horiz.index.TUM}
\alias{calc.field.horiz.index.TUM}
\title{Calculate horizontal structure index}
\usage{
calc.field.horiz.index.TUM(H.vec, D.vec, area, max.SDI = 700)
}
\arguments{
\item{H.vec}{Vector of tree heights in m}

\item{D.vec}{Vector of tree DBH in cm}

\item{area}{...of the plot in m2}

\item{max.SDI}{Maximum possible stand density index (SDI) (Default=700)}
}
\value{
Index of horizontal forest structure
}
\description{
# Function for horizontal index from TUM calculated from inventory data.
}
\examples{
in progress
}
\keyword{forest}
\keyword{heterogeneity}
\keyword{horizontal}
\keyword{structure}

