% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.field.vert.index.TUM_function.R
\name{calc.field.vert.index.TUM}
\alias{calc.field.vert.index.TUM}
\title{Calculate vertical structure index}
\usage{
calc.field.vert.index.TUM(H.vec, D.vec)
}
\arguments{
\item{H.vec}{Vector of tree heights in m}

\item{D.vec}{Vector of tree DBH in cm}
}
\value{
Index of vertical forest structure
}
\description{
Function for vertical index from TUM calculated from inventory data
}
\examples{
in progress
}
\keyword{forest}
\keyword{heterogeneity}
\keyword{structure}
\keyword{vertical}
