% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.gaussian.function.2D_function.R
\name{calc.gaussian.function.2D}
\alias{calc.gaussian.function.2D}
\title{Calculate Gaussian function value for a point in 2D space}
\usage{
calc.gaussian.function.2D(Xpt, Ypt, amp = 1, Xctr = 0, Yctr = 0, sd = 1)
}
\arguments{
\item{Xpt}{X-coordinate of the point}

\item{Ypt}{Y-coordinate of the point}

\item{amp}{Amplitude of the peak of the Gaussian curve}

\item{Xctr}{X-coordinate of the center point}

\item{Yctr}{Y-coordinate of the center point}

\item{sd}{Standard deviation of the Gaussian curve (for a curve that contains >99\% of the area between -1 and 1 use sd=0.4)}
}
\value{
Vector of function values for each input point
}
\description{
Function that returns for each X-Y-coordinate pair the value of a Gaussian function.
}
\examples{
in progress
}
\keyword{2D}
\keyword{Gaussian}
\keyword{bell}
\keyword{curve}
\keyword{density}
\keyword{distribution}
\keyword{normal}
\keyword{weighting}
