% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.gaussian.function_function.R
\name{calc.gaussian.function}
\alias{calc.gaussian.function}
\title{Calculate Gaussian function value for a point}
\usage{
calc.gaussian.function(pt, amp = 1, ctr = 0, sd = 1)
}
\arguments{
\item{pt}{Coordinate of the point}

\item{amp}{Amplitude of the peak of the Gaussian curve}

\item{ctr}{Coordinate of the center point}

\item{sd}{Standard deviation of the Gaussian curve (for a curve that contains >99\% of the area between -1 and 1 use sd=0.4)}
}
\value{
Vector of function values for each input point
}
\description{
Function that returns for each given point the value of a Gaussian function.
}
\examples{
in progress
}
\keyword{Gaussian}
\keyword{bell}
\keyword{curve}
\keyword{density}
\keyword{distribution}
\keyword{normal}
\keyword{weighting}

