% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lidar.horiz.index.DLR.from.prof_function.R
\name{calc.lidar.horiz.index.DLR.from.prof}
\alias{calc.lidar.horiz.index.DLR.from.prof}
\title{Calculate horizontal structure index from Lidar profile}
\usage{
calc.lidar.horiz.index.DLR.from.prof(prof, global.h.max = 50)
}
\arguments{
\item{prof}{Lidar profile as vector with height bins as names}

\item{global.h.max}{Maximal tree height in the region for normalization (Default=50)}
}
\value{
Index of horizontal forest structure
}
\description{
Function for horizontal index from DLR calculated from Lidar
vertical profile data
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{forest}
\keyword{heterogeneity}
\keyword{horizontal}
\keyword{profile}
\keyword{structure}

