% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lidar.horiz.index.DLR.from.xyz_function.R
\name{calc.lidar.horiz.index.DLR.from.xyz}
\alias{calc.lidar.horiz.index.DLR.from.xyz}
\title{Calculate horizontal structure index from Lidar point cloud}
\usage{
calc.lidar.horiz.index.DLR.from.xyz(XYZtable, global.h.max = 50)
}
\arguments{
\item{XYZtable}{Lidar point cloud or CHM in XYZ-table format}

\item{global.h.max}{Maximal tree height in the region for normalization (Default=50)}
}
\value{
Index of horizontal forest structure
}
\description{
Function for horizontal index from DLR calculated from Lidar
XYZ-table data (point cloud or CHM)
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{XYZ}
\keyword{forest}
\keyword{heterogeneity}
\keyword{horizontal}
\keyword{structure}
