% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lidar.horiz.index.UFZ.from.xyz_function.R
\name{calc.lidar.horiz.index.UFZ.from.xyz}
\alias{calc.lidar.horiz.index.UFZ.from.xyz}
\title{Calculate horizontal structure index from Lidar point cloud}
\usage{
calc.lidar.horiz.index.UFZ.from.xyz(XYZtable)
}
\arguments{
\item{XYZtable}{Lidar point cloud or CHM in XYZ-table format}
}
\value{
Index of horizontal forest structure
}
\description{
Function for horizontal index from UFZ (inspired by TUM index for field data)
calculated from Lidar XYZ-table data (point cloud or CHM)
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{XYZ}
\keyword{forest}
\keyword{heterogeneity}
\keyword{horizontal}
\keyword{structure}
