% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lidar.vert.index.DLR.from.xyz_function.R
\name{calc.lidar.vert.index.DLR.from.xyz}
\alias{calc.lidar.vert.index.DLR.from.xyz}
\title{Calculate vertical structure index from Lidar point cloud}
\usage{
calc.lidar.vert.index.DLR.from.xyz(XYZtable)
}
\arguments{
\item{XYZtable}{Lidar point cloud or CHM in XYZ-table format}
}
\value{
Index of vertical forest structure
}
\description{
Function for vertical index from DLR calculated from Lidar
XYZ-table data (point cloud or CHM)
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{XYZ}
\keyword{forest}
\keyword{heterogeneity}
\keyword{structure}
\keyword{vertical}
