% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calc.lidar.vert.index.DLR2.rel.centr.mass.from.prof_function.R
\name{calc.lidar.vert.index.DLR2.rel.centr.mass.from.prof}
\alias{calc.lidar.vert.index.DLR2.rel.centr.mass.from.prof}
\title{Calculate vertical structure index suggestion of Victor Cazcarra (DLR)
from Lidar profile or vertical foliage profile}
\usage{
calc.lidar.vert.index.DLR2.rel.centr.mass.from.prof(
  prof,
  perc = 0.5,
  h.min = 1
)
}
\arguments{
\item{perc}{Fraction between 0 and 1 (default 0.5) to specify Hperc = perc * Hmax}

\item{h.min}{Minimum height below which Lidar returns are ignored}

\item{XYZtable}{Lidar point cloud or CHM in XYZ-table format}
}
\value{
Index of vertical forest structure
}
\description{
Function for vertical index from DLR calculated from Lidar
XYZ-table data (point cloud or CHM).
The equation for the index V is: V = 1 - (abs(CM - Hperc) / Hperc) with
CM being the center of mass (equal to MCH or TCH, depending if input is
point cloud or CHM-raster) and Hperc being a custom fraction of the max.
height (Hmax) in the plot (default 50 percent).
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{XYZ}
\keyword{forest}
\keyword{heterogeneity}
\keyword{structure}
\keyword{vertical}
