% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lidar.vert.index.UFZ.from.prof_function.R
\name{calc.lidar.vert.index.UFZ.from.prof}
\alias{calc.lidar.vert.index.UFZ.from.prof}
\title{Calculate vertical structure index from Lidar profile}
\usage{
calc.lidar.vert.index.UFZ.from.prof(prof)
}
\arguments{
\item{prof}{Lidar profile as vector with height bins as names}
}
\value{
Index of vertical forest structure
}
\description{
Function for horizontal index from UFZ (inspired by TUM index for field data) 
calculated from Lidar vertical profile data
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{forest}
\keyword{heterogeneity}
\keyword{profile}
\keyword{structure}
\keyword{vertical}
