% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.P.H.ratio_function.R
\name{calc.P.H.ratio}
\alias{calc.P.H.ratio}
\title{Calculate P:H ratio}
\usage{
calc.P.H.ratio(prof)
}
\arguments{
\item{prof}{Lidar profile as vector with height bins as names}
}
\value{
number P:H
}
\description{
Function that calculates the ratio P:H of a canopy profile (Marvin et al. 2014),
where P is the height of the maximal peak in the profile (highest return
or foliage density) and H is the maximal height of the profile.
}
\examples{
in progress
}
\keyword{CHM}
\keyword{cloud}
\keyword{height}
\keyword{lidar}
\keyword{point}
\keyword{profile}
\keyword{xyz}
