% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.Shannon.index_function.R
\name{calc.ShannonIndex}
\alias{calc.ShannonIndex}
\title{Calculate the Shannon index of a profile}
\usage{
calc.ShannonIndex(vec, ex.vec = NA)
}
\arguments{
\item{vec}{Vector of counts or relative proportions}

\item{ex.vec}{Vector indices of elements in vec that should be excluded 
(e.g. to exclude ground return of a Lidar profile set to 1)}
}
\value{
Shannon index
}
\description{
Function that returns the Shannon diversity index of a vector
of counts or relative proportions (e.g. Lidar profile).
}
\examples{
in progress
}
\keyword{Lidar}
\keyword{Shannon}
\keyword{diversity}
\keyword{height}
\keyword{heterogeneity}
\keyword{index}
\keyword{profile}

