% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.skewness.R
\name{calc.skewness}
\alias{calc.skewness}
\title{Calculate Skewness}
\usage{
calc.skewness(vec)
}
\arguments{
\item{vec}{Vector of data}
}
\value{
Skewness (negative = left-skewed / positive = right-skewed)
}
\description{
Function for calculation of skewness of an empirical distribution
}
\examples{
calc.skewness(c(1,2,3,3,3,4))
}
\author{
Nikolai Knapp
}
\keyword{distribution}
\keyword{mean}
\keyword{profile}
\keyword{shape}
\keyword{skewness}
\keyword{symmetry}

