% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.TCH.from.point.cloud_function.R
\name{calc.TCH.from.point.cloud}
\alias{calc.TCH.from.point.cloud}
\title{Calculate mean top-of-canopy height (TCH) from XYZ-dataframe}
\usage{
calc.TCH.from.point.cloud(XYZdf, pxl.res = 1)
}
\arguments{
\item{XYZdf}{Point cloud in XYZ-dataframe format}

\item{pxl.res}{Side length of CHM pixel}
}
\value{
mean top-of-canopy height (TCH)
}
\description{
Function to calculate TCH from the point cloud in XYZ-dataframe format for a custom pixel size.
TCH is the mean of all CHM pixel values. So within each pixel the highest
return is taken and then the average height over all pixels is calculated.
}
\examples{
in progress
}
\keyword{CHM}
\keyword{Lidar}
\keyword{TCH}
\keyword{XYZ}
\keyword{forest}
\keyword{height}
\keyword{mean}
\keyword{pixel}
\keyword{profile}
\keyword{raster}
\keyword{resolution}
\keyword{scale}
\keyword{top-of-canopy}
