% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.weighted.mean_function.R
\name{calc.weighted.mean}
\alias{calc.weighted.mean}
\title{Calculate weighted mean}
\usage{
calc.weighted.mean(val, wt)
}
\arguments{
\item{val}{Vector of values}

\item{wt}{Vector of weights (e.g. counts or relative proportions)}
}
\value{
Mean
}
\description{
Function for weighted mean calculation in case of already
aggregated data (e.g. Lidar profile)
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{deviation}
\keyword{mean}
\keyword{standard}
\keyword{variance}
\keyword{weighted}

