% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.weighted.quantile_function.R
\name{calc.weighted.quantile}
\alias{calc.weighted.quantile}
\title{Calculate weighted quantile}
\usage{
calc.weighted.quantile(val, wt, quant)
}
\arguments{
\item{val}{Vector of values}

\item{wt}{Vector of weights (e.g. counts or relative proportions)}

\item{quant}{Quantile as float between 0 and 1}
}
\value{
Quantile
}
\description{
Function for weighted quantile calculation in case of already
aggregated data (e.g. Lidar profile)
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{percentile}
\keyword{quantile}
\keyword{weighted}
