% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.weighted.SD_function.R
\name{calc.weighted.SD}
\alias{calc.weighted.SD}
\title{Calculate weighted standard deviation}
\usage{
calc.weighted.SD(val, wt)
}
\arguments{
\item{val}{Vector of values}

\item{wt}{Vector of weights (e.g. counts or relative proportions)}
}
\value{
Standard deviation
}
\description{
Function for weighted SD calculation in case of already
aggregated data (e.g. Lidar profile)
}
\examples{
val.vec <- c(3, 5, 4, 1)
wt.vec <- c(1, 1, 5, 5)
weighted.SD(val=val.vec, wt=wt.vec)
}
\keyword{SD}
\keyword{deviation}
\keyword{standard}
\keyword{variance}
\keyword{weighted}
