% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpareto_function.R
\name{ddpareto}
\alias{ddpareto}
\title{Calculate values of a discrete power law distribution}
\usage{
ddpareto(x, lambda)
}
\arguments{
\item{x}{Value of the independent variable}

\item{lambda}{Scaling parameter}
}
\value{
Value of the dependent variable
}
\description{
Function that returns values of a discrete Pareto (power law)
distribution density (e.g. canopy gap sizes, tree diameters)
for a given scaling parameter lambda;
Code adapted from Asner et al. (2013) supplements.
}
\examples{
in progress
}
\keyword{Pareto}
\keyword{discrete}
\keyword{distribution}
\keyword{lambda}
\keyword{law}
\keyword{power}
\keyword{size}
