% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec.deg.to.deg.min.sec.R
\name{dec.deg.to.deg.min.sec}
\alias{dec.deg.to.deg.min.sec}
\title{Convert coordinates from decimal degrees to degrees, minutes and seconds}
\usage{
dec.deg.to.deg.min.sec(dd)
}
\arguments{
\item{dd}{Decimal degrees}
}
\value{
Degrees, minutes and seconds
}
\description{
Function that converts between coordinate notations
}
\examples{
dec.deg.to.deg.min.sec(10.51)
}
\keyword{coordinate}
\keyword{decimal}
\keyword{degree}
\keyword{minute}
\keyword{second}

