% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.min.sec.to.dec.deg.R
\name{deg.min.sec.to.dec.deg}
\alias{deg.min.sec.to.dec.deg}
\title{Convert coordinates from degrees, minutes and seconds to decimal degrees}
\usage{
deg.min.sec.to.dec.deg(deg, min, sec)
}
\arguments{
\item{deg}{Degrees}

\item{min}{Minutes}

\item{sec}{Seconds}
}
\value{
Decimal degrees
}
\description{
Function that converts between coordinate notations
}
\examples{
deg.min.sec.to.dec.deg(5, 4, 3)
}
\keyword{coordinate}
\keyword{decimal}
\keyword{degree}
\keyword{minute}
\keyword{second}
