% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_point_cloud_function.R
\name{display_point_cloud}
\alias{display_point_cloud}
\title{Plot point cloud in 3D space}
\usage{
display_point_cloud(
  xyz.dt,
  col.palette = NA,
  col.var = "Z",
  col.lim = NA,
  size = 1
)
}
\arguments{
\item{col.palette}{Color palette}

\item{col.var}{String with name of the column which contains the information for coloring}

\item{col.lim}{Vector of two elements specifying the limits of the col.var values}

\item{size}{...of plotted points}

\item{pc}{Point cloud in XYZ-table format (important: column names need to be X, Y and Z)}
}
\value{
3D plot in rgl window
}
\description{
Function for 3-dimensional point cloud visualization.
}
\author{
Nikolai Knapp
}
\keyword{3D}
\keyword{cloud}
\keyword{display}
\keyword{graphics}
\keyword{lidar}
\keyword{plot}
\keyword{point}
\keyword{visualization}
