% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.point.values.from.raster_function.R
\name{extract.point.values.from.raster}
\alias{extract.point.values.from.raster}
\title{Extract point values from raster}
\usage{
extract.point.values.from.raster(pc, ras)
}
\arguments{
\item{pc}{Input point cloud}

\item{ras}{Input raster}
}
\value{
Vector of raster values for each point, that can be added as a 
new column to the point cloud dataframe
}
\description{
Function that extracts values from an underlying raster at the position 
of each point of an overlying point cloud.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{DTM}
\keyword{cloud}
\keyword{extract}
\keyword{point}
\keyword{raster}

