% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrema.finder_function.R
\name{extrema.finder}
\alias{extrema.finder}
\title{Find extrema}
\usage{
extrema.finder(curve, wr, extr)
}
\value{

}
\description{
Function that finds local extrema in curves like canopy
height profiles with a moving window.
}
\examples{
in progress
}
