% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.nearest.neighbors_function.R
\name{find.nearest.neighbors}
\alias{find.nearest.neighbors}
\title{Find nearest neighbors in point cloud}
\usage{
find.nearest.neighbors(pc, n = 2)
}
\arguments{
\item{pc}{Point cloud as data.frame with at least a column "X" and optional columns "Y" and "Z"}

\item{n}{Number of neighbors to find (1 only returns the focal point itself)}
}
\value{
list with up to five elements which are all matrices where each row represents
each focal point in the point cloud and each column represents one neighbor sorted by
increasing distance. The matrix elements represent 1) neighbor indices, 2) neighbor distances,
3) neighbor X-coodinates, 4) neighbor Y-coodinates, 5) neighbor Z-coodinates.
}
\description{
Function that finds the n nearest neighbor points for each point
in a point cloud of up to 3 dimensions
}
\examples{
in progress
}
\keyword{ANN}
\keyword{coordinates}
\keyword{distance}
\keyword{find}
\keyword{nearest}
\keyword{neighbor}
\keyword{search}

