% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill_function.R
\name{gapfill}
\alias{gapfill}
\title{Fill raster gaps}
\usage{
gapfill(in.ras, fill.window.size = 5, fill.window.type = "circle",
  fill.value = "mean")
}
\arguments{
\item{in.ras}{Input raster}

\item{fill.window.size}{Radius of the window around the focal pixel from which a filling value is calculated}

\item{fill.window.type}{Shape of the window must be "circle" or "rectangle"}

\item{fill.value}{Replacing value, e.g. "mean", "min", "max", of the filling window or simply 0 or NA}
}
\value{
Filled raster object
}
\description{
Function that fills gaps in a raster.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{NA}
\keyword{fill}
\keyword{gap}
\keyword{hole}
\keyword{raster}
\keyword{rasterization}
\keyword{replace}
\keyword{window}

