% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.cuboid_function.R
\name{in.cuboid}
\alias{in.cuboid}
\title{Check if point lies within a cuboid}
\usage{
in.cuboid(Xcor, Ycor, Zcor, Xctr, Yctr, Zbase, sidelength, height)
}
\arguments{
\item{Xcor}{X-coordinate of the point}

\item{Ycor}{Y-coordinate of the point}

\item{Zcor}{Z-coordinate of the point}

\item{Xctr}{X-coordinate of the center}

\item{Yctr}{Y-coordinate of the center}

\item{Zbase}{Z-coordinate of the lower end}

\item{sidelength}{horizontal length and width of the cuboid}

\item{height}{...of the cuboid}
}
\value{
boolean or matrix of booleans
}
\description{
Function that checks for each X-Y-coordinate pair, whether the point lies in a
cuboid of a given center, side length and height. Works with single numbers or matrices of
X- and Y-coordinates as input.
}
\examples{
in progress
}
\keyword{cuboid}
\keyword{in}
\keyword{shape}

