% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.icecone_function.R
\name{in.icecone}
\alias{in.icecone}
\title{Check if point lies within an "ice cone" like shape}
\usage{
in.icecone(Xcor, Ycor, Zcor, Xctr, Yctr, Zctr, radius, height)
}
\arguments{
\item{Xcor}{X-coordinate of the point}

\item{Ycor}{Y-coordinate of the point}

\item{Zcor}{Z-coordinate of the point}

\item{Xctr}{X-coordinate of the center}

\item{Yctr}{Y-coordinate of the center}

\item{Zctr}{Z-coordinate of the center}

\item{radius}{...of the structure}

\item{height}{...of the structure}
}
\value{
boolean or matrix of booleans
}
\description{
Function that checks for each X-Y-coordinate pair, whether the point lies in an
"ice cone" of a given center, radius and height. The ice cone consists of a hemi-
spheroid forming the upper half (ice ball) and a bottom-up cone forming the 
lower half (waffle).
Works with single numbers or matrices of X- and Y-coordinates as input.
}
\examples{
in progress
}
\keyword{cone}
\keyword{crown}
\keyword{ice}
\keyword{in}
\keyword{shape}
\keyword{spheroid}
\keyword{tree}

