% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_circle_function.R
\name{in_circle}
\alias{in_circle}
\title{Check if point lies within a circle}
\usage{
in_circle(Xcor, Ycor, Xctr, Yctr, radius)
}
\arguments{
\item{Xcor}{X-coordinate of the point}

\item{Ycor}{Y-coordinate of the point}

\item{Xctr}{X-coordinate of the circle center}

\item{Yctr}{Y-coordinate of the circle center}

\item{radius}{...of the circle}
}
\value{
boolean or matrix of booleans
}
\description{
Function that checks for each X-Y-coordinate pair, whether the point lies in a
circle of a given center and radius. Works with single numbers or matrices of
X- and Y-coordinates as input.
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{circle}
\keyword{in}
\keyword{overlap}
\keyword{overlay}
