% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_spheroid_function.R
\name{in_spheroid}
\alias{in_spheroid}
\title{Check if point lies within a spheroid}
\usage{
in_spheroid(Xcor, Ycor, Zcor, Xctr, Yctr, Zctr, radius, height)
}
\arguments{
\item{Xcor}{X-coordinate of the point}

\item{Ycor}{Y-coordinate of the point}

\item{Zcor}{Z-coordinate of the point}

\item{Xctr}{X-coordinate of the center}

\item{Yctr}{Y-coordinate of the center}

\item{Zctr}{Z-coordinate of the center}

\item{radius}{...of the equator plain of the spheroid}

\item{height}{...of the vertical axis of the spheroid}
}
\value{
boolean or matrix of booleans
}
\description{
Function that checks for each X-Y-coordinate pair, whether the point lies in a
spheroid of a given center, radius and height. Works with single numbers or matrices of
X- and Y-coordinates as input.
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{in}
\keyword{shape}
\keyword{spheroid}
