% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_area_density_function.R
\name{leaf_area_density}
\alias{leaf_area_density}
\title{Reconstruct leaf area density (LAD) from lidar point clouds}
\usage{
leaf_area_density(
  pc,
  xy.res = 1,
  z.res = 1,
  k = 1,
  normalized = F,
  ground.class = 2
)
}
\arguments{
\item{pc}{Point cloud in data.table, data.frame or LAS format (important: column names need to be X, Y and Z)}

\item{xy.res}{Side length of a voxel in the horizontal plain}

\item{z.res}{Side length of a voxel in vertical direction}

\item{k}{Conversion coefficient (light extinction coefficient)}

\item{normalized}{Boolean whether the point cloud is terrain normalized or
not. For normalized point clouds all points of height < 0.5 m are considered
ground points. For non-normalized point clouds a column "Classification" is
required and ground points should be labelled.}

\item{ground.class}{Only required for non-normalized point clouds. It
provides the numeric code of ground points in the "Classification" column.
The standard value is 2.}
}
\value{
data.table with each row representing one voxel. Coordinates
represent voxel centers and LAD and cumulative LAI are given for each voxel
}
\description{
Function to reconstruct leaf area density (LAD) per volume based on the
MacArthur-Horn equation, which inverts the Beer-Lambert principle of
light extinction. It provides an estimate of LAD per voxel.
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
