% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.circular.moving.window.matrix_function.R
\name{make.circular.moving.window.matrix}
\alias{make.circular.moving.window.matrix}
\title{Create a circular moving window matrix}
\usage{
make.circular.moving.window.matrix(rad = 10)
}
\arguments{
\item{rad}{Radius of the considered (circular) neighborhood}
}
\value{
matrix with 1 in circle and NA outside the circle
}
\description{
Function that creates a circular moving window matrix.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{circle}
\keyword{filter}
\keyword{focal}
\keyword{moving}
\keyword{raster}
\keyword{window}

