% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.large.footprint.lidar.pulse_function.R
\name{make.large.footprint.lidar.pulse}
\alias{make.large.footprint.lidar.pulse}
\title{Scan a voxel forest with a virtual large footprint Lidar}
\usage{
make.large.footprint.lidar.pulse(vxf, Xctr, Yctr, diameter = 25,
  binwidth = 1, k = 0.2, sd = 0.4, VG.ratio = 2.5)
}
\arguments{
\item{vxf}{voxel forest created with make.voxelforest function}

\item{Xctr}{X-coordinate of the pulse center}

\item{Yctr}{Y-coordinate of the pulse center}

\item{diameter}{Diameter of the pulse footprint in meters (e.g. LVIS = 25 m, Icesat GLAS = 70 m)}

\item{binwidth}{Vertical resolution of the waveform in meters}

\item{k}{Extinction coefficient}

\item{sd}{Standard deviation of the Gaussian energy distribution (for a curve that approaches 0 at the margin of the pulse use sd = 0.4*radius)}

\item{VG.ratio}{Reflectance ratio between vegetation surface and ground surface}
}
\value{
Vector of the Lidar waveform (normalized to a sum of 1)
}
\description{
Function that simulates a Lidar waveform of a given voxel forest and returns a profile vector.
The horizontal energy distribution inside the pulse is simulated with a Gaussian distribution.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{footprint}
\keyword{forest}
\keyword{gaussian}
\keyword{large}
\keyword{lidar}
\keyword{plot}
\keyword{profile}
\keyword{pulse}
\keyword{voxel}
\keyword{waveform}

