% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.lidarscan.dt_function.R
\name{make.lidarscan.dt}
\alias{make.lidarscan.dt}
\title{Scan a voxel forest with a virtual Lidar}
\usage{
make.lidarscan.dt(vxf.dt, P0.AGR = 0.2, k.AGR = 0.2, P0.GR = 0.2,
  k.GR = 0.2, LAD.prefactor = F, prob.out = F)
}
\arguments{
\item{vxf.dt}{data.table containing the voxel forest (derived from make.voxelforest.dt function)}

\item{P0.AGR}{Initial return probability for above ground voxels}

\item{k.AGR}{Extinction coefficient for above ground voxels}

\item{P0.GR}{Initial return probability for ground voxels}

\item{k.GR}{Extinction coefficient for ground voxels}

\item{LAD.prefactor}{Boolean to specify whether leaf area density
(LAD) of a voxel should be multiplied with P0 to get final return
probability or not}

\item{prob.out}{Boolean to specify whether an additional output containing the
full voxelforest with information on canopy above and Lidar return
probability for each voxel should be produced}
}
\value{
data.table object containing a XYZ-table of discrete Lidar returns (point cloud).
If prob.out=T the function will return a list containing two data.tables: 1) the Lidar
point cloud and 2) the voxel forest with Lidar return probability info.
}
\description{
Function that simulates a Lidar point cloud of a given voxel forest and writes it to a file in XYZ-table format.
The model uses a Beer-Lambert light extinction approach and interprets it as probability
to get a Lidar return from a voxel.
}
\examples{
in progress
}
\keyword{cloud}
\keyword{forest}
\keyword{lidar}
\keyword{plot}
\keyword{point}
\keyword{stand}
\keyword{voxel}
\keyword{xyz}

