% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.map.from.spatial.index_function.R
\name{make.map.from.spatial.index}
\alias{make.map.from.spatial.index}
\title{Make rasters from values and spatial indices}
\usage{
make.map.from.spatial.index(
  spatial.ID,
  value,
  res = 1,
  minx = NA,
  miny = NA,
  maxx = NA,
  maxy = NA
)
}
\arguments{
\item{spatial.ID}{vector containing spatial indices}

\item{value}{vector containing values to fill the pixels}

\item{res}{side length of one spatial subunit}

\item{minx}{minimum X-coordinate}

\item{miny}{minimum Y-coordinate}

\item{maxx}{maximum X-coordinate}

\item{maxy}{maximum Y-coordinate}
}
\value{
raster object
}
\description{
Function that returns a raster based on given spatial indices (plot numbers)
and corresponding values that should be assigned to each pixel.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{cast}
\keyword{index}
\keyword{map}
\keyword{number}
\keyword{plot}
\keyword{raster}
\keyword{spatial}
\keyword{value}
