% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.polygon_function.R
\name{make.polygon}
\alias{make.polygon}
\title{Make a polygon}
\usage{
make.polygon(vertex.mx)
}
\arguments{
\item{vertex.mx}{Matrix with two columns (X- and Y-coordinates) where each row represents one vertex}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Function that creates a spatial polygon data frame from given vertex coordinates.
}
\examples{
test.spdf <- make.polygon(vertex.mx=matrix(c(1,1,5,1,3,3), ncol=2, byrow=T))
plot(test.spdf)
}
\keyword{area}
\keyword{polygon}
\keyword{shape}
\keyword{vector}
\keyword{vertex}
