% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.profile.from.XYZ.value_function.R
\name{make.profile.from.XYZ.value}
\alias{make.profile.from.XYZ.value}
\title{Derive a vertical profile of a point cloud for a certain variable}
\usage{
make.profile.from.XYZ.value(XYZ.value.table, binwidth = 1, stat = "sum")
}
\arguments{
\item{XYZ.value.table}{4-column data frame with 3rd column containing heights and 4th column containing values for which to make the profile}

\item{binwidth}{number specifying the vertical width of height classes}

\item{stat}{desired statistic ("sum", "count", "max", "min", "mean", "median")}
}
\value{
vector of point or pixel counts per height class with height values as names
}
\description{
Function that creates profiles of a certain statistic of a certain variable
from XYZ-tables with a custom height resolution (binwidth)
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{cloud}
\keyword{lidar}
\keyword{point}
\keyword{profile}
\keyword{xyz}

