% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.profile.from.Z_function.R
\name{make.profile.from.Z}
\alias{make.profile.from.Z}
\title{Derive a vertical profile of a vector of heights}
\usage{
make.profile.from.Z(Z.vec, binwidth = 1)
}
\arguments{
\item{Z.vec}{vector containing single height measurements}

\item{binwidth}{number specifying the vertical width of height classes}
}
\value{
vector of point or pixel counts per height class with height values as names
}
\description{
Function that creates profiles from Z-vectors with a custom height
resolution (binwidth)
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{Z}
\keyword{cloud}
\keyword{height}
\keyword{lidar}
\keyword{point}
\keyword{profile}
\keyword{vector}
\keyword{xyz}
