% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.spatial.index.matrix_function.R
\name{make.spatial.index.matrix}
\alias{make.spatial.index.matrix}
\title{Make a spatial index matrix}
\usage{
make.spatial.index.matrix(minx = 0, maxx, miny = 0, maxy, res)
}
\arguments{
\item{minx}{minimal X-coordinate}

\item{maxx}{maximal X-coordinate}

\item{miny}{minimal Y-coordinate}

\item{maxy}{maximal Y-coordinate}

\item{res}{side length of one spatial subunit}
}
\value{
matrix of the arrangement of spatial indices
}
\description{
Create a sampling plot matrix for a given area with a 
specific resolution
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{index}
\keyword{number}
\keyword{plot}
\keyword{spatial}

