% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.vectorgrid_function.R
\name{make.vectorgrid}
\alias{make.vectorgrid}
\title{Make a vectorgrid}
\usage{
make.vectorgrid(
  ll.x = 0,
  ll.y = 0,
  size.x = 1,
  size.y = 1,
  n.x = 1,
  n.y = 1,
  rot.angle = 0
)
}
\arguments{
\item{ll.x}{X-coordinate of the lower left corner}

\item{ll.y}{Y-coordinate of the lower left corner}

\item{size.x}{Side length of one cell in X-direction}

\item{size.y}{Side length of one cell in Y-direction}

\item{n.x}{Number of cells in X-direction}

\item{n.y}{Number of cells in Y-direction}

\item{rot.angle}{Rotation angle clockwise in degrees}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Function that creates as vectorgrid of rectangular cells.
}
\examples{
testgrid.spdf <- make.vectorgrid(ll.x=0, ll.y=0, size.x=2, size.y=3, n.x=10, n.y=5)
plot(testgrid.spdf)
sub.spdf <- subset(testgrid, ID==1)
plot(sub.spdf, add=T, col="red")
sub.spdf <- subset(testgrid, ID==2)
plot(sub.spdf, add=T, col="green")
sub.spdf <- subset(testgrid, ID==11)
plot(sub.spdf, add=T, col="blue")
}
\keyword{fishnet}
\keyword{grid}
\keyword{lattice}
\keyword{matrix}
\keyword{polygon}
\keyword{sampling}
\keyword{shape}
\keyword{vector}
