% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.ycor.matrix_function.R
\name{make.ycor.matrix}
\alias{make.ycor.matrix}
\title{Make a Y-coordinate matrix}
\usage{
make.ycor.matrix(minx = 0, maxx, miny = 0, maxy, res = 1)
}
\arguments{
\item{minx}{minimal X-coordinate}

\item{maxx}{maximal X-coordinate}

\item{miny}{minimal Y-coordinate}

\item{maxy}{maximal Y-coordinate}

\item{res}{side length of one cell}
}
\value{
matrix of Y-coordinates
}
\description{
Create a matrix that contains the Y-coordinate of each cell as cell value
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{Y}
\keyword{coordinate}
\keyword{matrix}

