% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.zcor.array_function.R
\name{make.zcor.array}
\alias{make.zcor.array}
\title{Make a Z-coordinate array}
\usage{
make.zcor.array(minx = 0, maxx, miny = 0, maxy, minz = 0, maxz, res = 1)
}
\arguments{
\item{minx}{minimal X-coordinate}

\item{maxx}{maximal X-coordinate}

\item{miny}{minimal Y-coordinate}

\item{maxy}{maximal Y-coordinate}

\item{minz}{minimal Z-coordinate}

\item{maxz}{maximal Z-coordinate}

\item{res}{side length of one cell}
}
\value{
3D array of Z-coordinates
}
\description{
Create a 3D array that contains the Z-coordinate of each cell as cell value
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{3D}
\keyword{Z}
\keyword{array}
\keyword{coordinate}
\keyword{voxel}

