% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_color_vector_function.R
\name{make_color_vector}
\alias{make_color_vector}
\title{Make a color vector}
\usage{
make_color_vector(vec, palette = NA, min = NA, max = NA, step = 1, alpha = 1)
}
\arguments{
\item{vec}{Value vector}

\item{min}{Minimum value to consider for the scaling of the color values (may subceed the value vector)}

\item{max}{Maximum value to consider for the scaling of the color values (may exceed the value vector)}

\item{step}{Step width between values to consider for the scaling of the color values}

\item{alpha}{Transparency from 0 = fully transparent to 1 = fully opaque}

\item{col.palette}{Color palette, e.g. default R palettes or created with colorRampPalette function}
}
\value{
Vector of color codings with the same length and order as the input vector
}
\description{
Function that takes a vector of values and a desired color palette and returns
color values that correspond to each vector value.
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{coding}
\keyword{colors}
\keyword{palette}
\keyword{ramp}
\keyword{scale}
\keyword{values}
