% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_large_footprint_lidar_pulse_function.R
\name{make_large_footprint_lidar_pulse}
\alias{make_large_footprint_lidar_pulse}
\title{Scan a voxel forest with a virtual large footprint Lidar}
\usage{
make_large_footprint_lidar_pulse(
  xyz.dt,
  input = "vxf",
  Xctr,
  Yctr,
  diameter = 25,
  binwidth = 1,
  k = 0.2,
  sd = 0.5 * 12.5,
  VG.ratio = 2.5,
  weight.var = NA
)
}
\arguments{
\item{xyz.dt}{Input data.table which can either be a) a voxel forest created with make.voxelforest function (containing columns X, Y, Z and LAI among others) or b) a discrete lidar point cloud or CHM (containing columns X, Y and Z)}

\item{input}{Switch for the input type, which has to be either "vxf" for voxel forest or "lid" for discrete lidar data}

\item{Xctr}{X-coordinate of the pulse center}

\item{Yctr}{Y-coordinate of the pulse center}

\item{diameter}{Diameter of the pulse footprint in meters (e.g. LVIS = 25 m, SLICER = 10 m, Icesat GLAS = 70 m, GEDI = 25 m)}

\item{binwidth}{Vertical resolution of the waveform in meters}

\item{k}{Extinction coefficient (only relevant if input = "vxf")}

\item{sd}{Standard deviation of the Gaussian energy distribution (for a curve that approaches 0 at the margin of the pulse use sd = 0.4\*radius; for GEDI simulations commonly sd = 0.5\*radius is used, which leads to a more truncated energy distribution)}

\item{VG.ratio}{Reflectance ratio between vegetation surface and ground surface}

\item{weight.var}{Weighting variable, e.g. set to "Intensity" to use the intensity column in a lidar point cloud}
}
\value{
Vector of the Lidar waveform (normalized to a sum of 1)
}
\description{
Function that simulates a Lidar waveform of a) a given voxel forest or b) a given lidar point
cloud or CHM and returns a profile vector.
The horizontal energy distribution inside the pulse is simulated with a Gaussian distribution.
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{footprint}
\keyword{forest}
\keyword{gaussian}
\keyword{large}
\keyword{lidar}
\keyword{plot}
\keyword{profile}
\keyword{pulse}
\keyword{simulation}
\keyword{voxel}
\keyword{waveform}
