% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_profile_from_XYZ_function.R
\name{make_profile_from_XYZ}
\alias{make_profile_from_XYZ}
\title{Derive a vertical profile of a point cloud or a canopy height model}
\usage{
make_profile_from_XYZ(XYZ.table, binwidth = 1)
}
\arguments{
\item{XYZ.table}{3-column data frame with 3rd column containing heights}

\item{binwidth}{number specifying the vertical width of height classes}
}
\value{
vector of point or pixel counts per height class with height values as names
}
\description{
Function that creates profiles from XYZ-tables with a custom height
resolution (binwidth)
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{CHM}
\keyword{cloud}
\keyword{lidar}
\keyword{point}
\keyword{profile}
\keyword{xyz}
