% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic.boundaries.dt_function.R
\name{periodic.boundaries.dt}
\alias{periodic.boundaries.dt}
\title{Periodic boundaries}
\usage{
periodic.boundaries.dt(xyz.dt, minx, maxx, miny, maxy)
}
\arguments{
\item{xyz.dt}{Data.table containing X-, Y- and Z-coordinates of the points}

\item{minx}{Minimum X-coordinate of the area}

\item{maxx}{Maximum X-coordinate of the area}

\item{miny}{Minimum Y-coordinate of the area}

\item{maxy}{Maximum Y-coordinate of the area}
}
\value{
data.table of X-, Y- and Z-coordinates with periodic boundaries
}
\description{
Function that creates periodic boundaries for point clouds, i.e. points
that exceed the boundaries of the area re-appear on the opposite side 
of the area.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{border}
\keyword{boundaries}
\keyword{closed}
\keyword{cloud}
\keyword{crowns}
\keyword{edge}
\keyword{effects}
\keyword{periodic}
\keyword{point}
\keyword{tree}
\keyword{world}
\keyword{wrapped}
