% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic.boundaries.for.circular.plot_function.R
\name{periodic.boundaries.for.circular.plot}
\alias{periodic.boundaries.for.circular.plot}
\title{Periodic boundaries for circular forest plots}
\usage{
periodic.boundaries.for.circular.plot(vxf.dt, Xctr, Yctr, radius, clip = T)
}
\arguments{
\item{vxf.dt}{Data.table containing voxelforest object created with make.voxelforest; columns of the stem positions Xstem and Ystem are crucial}

\item{Xctr}{X-coordinate of the plot center}

\item{Yctr}{Y-coordinate of the plot center}

\item{radius}{of the circular plot}

\item{clip}{Boolean to specify whether output should be clipped to the circle (default) or all}
}
\value{
data.table containing the voxelforest with periodic boundaries
}
\description{
Function that creates periodic boundaries for voxelforests along circular plot borders
(e.g. large lidar footprints or circular inventory plots). Tree crown voxels
that exceed the boundaries of the circle re-appear on the opposite side
of the circle. The placement of trees is based on the polar coordinates of
the stem positions (distance and angle to the circle center). All trees with
stems inside the circle are copied to exist a second time outside of the
circle at the position opposite of the circle center and with the same distance
to the boundary, but outwards.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{border}
\keyword{boundaries}
\keyword{circle}
\keyword{closed}
\keyword{cloud}
\keyword{crowns}
\keyword{edge}
\keyword{effects}
\keyword{footprint}
\keyword{large}
\keyword{periodic}
\keyword{plot}
\keyword{point}
\keyword{tree}
\keyword{voxelforest}
\keyword{world}
\keyword{wrapped}

