% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic_raster_frame_function.R
\name{periodic_raster_frame}
\alias{periodic_raster_frame}
\title{Periodic raster frame}
\usage{
periodic_raster_frame(ras, w = 10)
}
\arguments{
\item{ras}{Raster object}

\item{w}{Width of the frame (number of pixels)}
}
\value{
raster object
}
\description{
Function that increases a raster by adding a frame around in which the
pixel values correspond to the values found at the opposite side of the
raster, creating periodic boundaries, e.g. for moving window analysis.
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{boundaries}
\keyword{buffer}
\keyword{edge}
\keyword{effect}
\keyword{filter}
\keyword{frame}
\keyword{margin}
\keyword{moving}
\keyword{periodic}
\keyword{raster}
\keyword{texture}
\keyword{window}
