% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras2colmx_function.R
\name{ras2colmx}
\alias{ras2colmx}
\title{Convert raster to bycol-matrix}
\usage{
ras2colmx(ras)
}
\arguments{
\item{ras}{Input raster object}
}
\value{
Matrix (cell order bycol)
}
\description{
Convert a raster with origin in lower left corner to a matrix with origin 
in upper left corner and with cell order filling column after column 
from top to bottom (bycol)
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{bycol}
\keyword{conversion}
\keyword{coordinate}
\keyword{matrix}
\keyword{origin}
\keyword{raster}
\keyword{transposition}

