% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras2rowmx_function.R
\name{ras2rowmx}
\alias{ras2rowmx}
\title{Convert raster to byrow-matrix}
\usage{
ras2rowmx(ras)
}
\arguments{
\item{ras}{Input raster object}
}
\value{
Matrix (cell order byrow)
}
\description{
Convert a raster with origin in lower left corner to a matrix with origin 
in upper left corner and with cell order filling row after row 
from left to right (byrow)
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{byrow}
\keyword{conversion}
\keyword{coordinate}
\keyword{matrix}
\keyword{origin}
\keyword{raster}
\keyword{transposition}

