% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras2xyzdf_function.R
\name{ras2xyzdf}
\alias{ras2xyzdf}
\title{Convert raster to XYZ-data.frame}
\usage{
ras2xyzdf(ras)
}
\arguments{
\item{ras}{Input raster object}
}
\value{
data.frame
}
\description{
Convert a raster to a XYZ-data.frame with Z being the pixel values
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{XYZ}
\keyword{conversion}
\keyword{coordinate}
\keyword{data.frame}
\keyword{raster}
\keyword{table}

