% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras2xyzdt_function.R
\name{ras2xyzdt}
\alias{ras2xyzdt}
\title{Convert raster to XYZ-data.table}
\usage{
ras2xyzdt(ras)
}
\arguments{
\item{ras}{Input raster object}
}
\value{
data.table
}
\description{
Convert a raster to a XYZ-data.table with Z being the pixel values
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{XYZ}
\keyword{conversion}
\keyword{coordinate}
\keyword{data.table}
\keyword{raster}
\keyword{table}
