% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.from.point.cloud_function.R
\name{raster.from.point.cloud}
\alias{raster.from.point.cloud}
\title{Convert a point cloud to a raster}
\usage{
raster.from.point.cloud(pc, xcor = "X", ycor = "Y", var = "Z", res = 1,
  func = "max")
}
\arguments{
\item{pc}{Point cloud in XYZ-table format}

\item{xcor}{Variable name that contains the X-coordinates (Default "X")}

\item{ycor}{Variable name that contains the Y-coordinates (Default "Y")}

\item{var}{Variable name that contains the values that should be rasterized (Default "Z")}

\item{res}{Resolution (cell side length) of the output raster}

\item{func}{Function to be applied (either "max", "min", "mean", "count", "sum" or "mode")}
}
\value{
Raster object
}
\description{
Function that creates a raster from a point cloud by applying a certain function
in each cell, e.g. func=max returns a surface raster (CHM in case of vegetation),
min returns lowest point per cell, mean returns the mean value of all points in the
cell, count returns a point density raster, sum returns the sum of all point values
and mode returns the most common value (in case of several equally common values one of
them is taken randomly).
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{XYZ}
\keyword{cloud}
\keyword{point}
\keyword{raster}
\keyword{rasterization}

