% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.large.dt.files_function.R
\name{rbind.large.dt.files}
\alias{rbind.large.dt.files}
\title{Read large tables from files and rbind them into one huge data table}
\usage{
rbind.large.dt.files(path, pattern)
}
\arguments{
\item{path}{...to the directory with the input files}

\item{pattern}{...that is containded in all input filenames}
}
\value{
data table containing all rows of all input tables
}
\description{
Function that reads a group of equally structured tables from files
(e.g. point clouds in XYZ-format) and integrates them into one huge
data table
}
\examples{
in progress
}
\keyword{XYZ}
\keyword{big}
\keyword{data}
\keyword{file}
\keyword{large}
\keyword{list}
\keyword{rbind}
\keyword{table}

