% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.outliers.from.point.cloud.IQR_function.R
\name{remove.outliers.from.point.cloud.IQR}
\alias{remove.outliers.from.point.cloud.IQR}
\title{Remove outliers from point cloud}
\usage{
remove.outliers.from.point.cloud.IQR(pc, n = 50, low.perc = 0.25,
  high.perc = 0.75, fac = 1.5, high = T, low = T)
}
\arguments{
\item{pc}{Input point cloud (first 3 columns need to be "X", "Y" and "Z")}

\item{n}{Number of considered neighbor points (with lowest XY-distance to focal point)}

\item{low.perc}{Lower end of the desired interquantile range (fraction between 0 and 1)}

\item{high.perc}{Upper end of the desired interquantile range (fraction between 0 and 1)}

\item{fac}{Factor by which a Z-value needs to exceed or subceed the IQR of its neighbors to be considered an outlier}

\item{high}{bool if false high outliers will not be removed}

\item{low}{bool if false low outliers will not be removed}
}
\value{
Point cloud dataframe with rows from the input that were detected as outliers removed
}
\description{
Function that removes points from a point cloud if their Z-value exceeds
the given percentile of their neighborhood's Z-values by a certain factor.
}
\examples{
in progress
}
\keyword{IQR}
\keyword{clean}
\keyword{cloud}
\keyword{filter}
\keyword{outlier}
\keyword{percentile}
\keyword{point}

