% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.outliers.from.raster.IQR_function.R
\name{remove.outliers.from.raster.IQR}
\alias{remove.outliers.from.raster.IQR}
\title{Remove outliers from raster}
\usage{
remove.outliers.from.raster.IQR(ras, rad = 5, low.perc = 0.25,
  high.perc = 0.75, fac = 1.5, high = T, low = T)
}
\arguments{
\item{ras}{Input raster}

\item{rad}{Radius of the considered (circular) neighborhood}

\item{low.perc}{Lower end of the desired interquantile range (fraction between 0 and 1)}

\item{high.perc}{Upper end of the desired interquantile range (fraction between 0 and 1)}

\item{fac}{Factor by which a pixel value needs to exceed or subceed the IQR of its neighbors to be considered an outlier}

\item{high}{bool if false high outliers will not be removed}

\item{low}{bool if false low outliers will not be removed}
}
\value{
Raster object with gaps (NA) at the positions of former outliers
}
\description{
Function that sets value of a raster pixel to NA if it exceeds or subceeds
the interquantile range (IQR) of its neighbors by a certain factor.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{DTM}
\keyword{IQR}
\keyword{clean}
\keyword{filter}
\keyword{interquantile}
\keyword{outlier}
\keyword{percentile}
\keyword{quantile}
\keyword{quartile}
\keyword{range}
\keyword{raster}

