% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject.points_function.R
\name{reproject.points}
\alias{reproject.points}
\title{Reproject points in space from one coordinate reference system (CRS) to another}
\usage{
reproject.points(xcor, ycor, in.crs, out.crs)
}
\arguments{
\item{xcor}{Vector of X-coordinates}

\item{ycor}{Vector of Y-coordinates}

\item{in.crs}{Original CRS of the data}

\item{out.crs}{Target CRS to which the points should be projected}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Function that reprojects X- and Y-coordinates between different CRS.
}
\examples{
latlon <- CRS("+init=epsg:4326")
utm33n <- CRS("+init=epsg:32633")
reproject.points(xcor=c(0, 100), ycor=c(0, 100), in.crs=utm33n, out.crs=latlon)
}
\keyword{CRS}
\keyword{coordinate}
\keyword{geolocation}
\keyword{points}
\keyword{projection}
\keyword{reference}
\keyword{system}
