% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.points_function.R
\name{rotate.points}
\alias{rotate.points}
\title{Rotate points around X-, Y- or Z-axis}
\usage{
rotate.points(cor.df, rot.angle, rot.axis = "Z", center = c(0, 0, 0))
}
\arguments{
\item{cor.df}{Coordinate dataframe with columns containing X-, Y- and Z-coordinate (at least
two columns are required and names need to be "X", "Y" or "Z", respectively)}

\item{rot.angle}{Rotation angle in degrees (clockwise rotation)}

\item{rot.axis}{Rotation axis (default is "Z", can also be set to "X" or "Y")}

\item{center}{Vector of three elements specifying the  X-, Y- and Z-coordinate of the rotation center point}
}
\value{
Coordinate dataframe with X- and Y-coordinates after rotation
}
\description{
Function to rotate a given set of points or point cloud in a
plain around center point with a certain angle.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{angle}
\keyword{cloud}
\keyword{point}
\keyword{rotation}
\keyword{transform}

