% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowmx2ras_function.R
\name{rowmx2ras}
\alias{rowmx2ras}
\title{Convert byrow-matrix to raster}
\usage{
rowmx2ras(mx)
}
\arguments{
\item{mx}{Input matrix}
}
\value{
Raster object
}
\description{
Convert a matrix with origin in upper left corner and with cell order from
left to right (byrow) to a raster with origin in lower left corner
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{byrow}
\keyword{conversion}
\keyword{coordinate}
\keyword{matrix}
\keyword{origin}
\keyword{raster}
\keyword{transposition}

