% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol.tree.positions_function.R
\name{sobol.tree.positions}
\alias{sobol.tree.positions}
\title{Distribute trees evenly in the plot considering their sizes}
\usage{
sobol.tree.positions(treesize.vec, plotsize = 20)
}
\arguments{
\item{treesize.vec}{Vector of tree sizes (e.g. DBH, BA or height)}

\item{plotsize}{Sidelength of a square shaped plot [m]}
}
\value{
Data.frame with 3 columns containing X- and Y-coordinates and tree sizes
}
\description{
Function that distributes trees evenly (quasi randomly) in the plot starting
with the largest and positioning them one after another with decreasing size
using the Sobol algorithm. The sobol algorithm needs initialization before
the first application of the function (see example below).
}
\examples{
sobol.init <- sobol(1, dim=2, init=T)
pos <- sobol.tree.positions(c(3, 2, 1), plotsize=10)
plot(pos$Y ~ pos$X)
}
\author{
Nikolai Knapp
}
\keyword{Sobol}
\keyword{coordinates}
\keyword{distribution}
\keyword{points}
\keyword{position}
\keyword{pseudo}
\keyword{quasi}
\keyword{random}
\keyword{space}
\keyword{stem}
\keyword{tree}

