% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.under.voxelforest_function.R
\name{terrain.under.voxelforest}
\alias{terrain.under.voxelforest}
\title{Add terrain under a voxelforest}
\usage{
terrain.under.voxelforest(vxf.dt, terrain.dt)
}
\arguments{
\item{vxf.dt}{Input data.table which contains a voxel forest created with make.voxelforest function and the specific settings "keep=C("TreeID", "X", "Y")", i.e. each voxel has to contain the unique tree ID and the stem position}

\item{terrain.dt}{Input data.table which contains X-, Y- and Z-column representing a digital terrain model at 1-m resolution (e.g., can be obtained from a 1-m DTM raster with data.table(ras2xyzdf(dtm.ras)))}
}
\value{
Data.table containing a voxel forest with terrain
}
\description{
Function that shifts a voxelforest vertically to add underlying terrain elevation.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{CHM}
\keyword{DEM}
\keyword{DSM}
\keyword{DTM}
\keyword{altitude}
\keyword{cloud}
\keyword{elevation}
\keyword{forest}
\keyword{lidar}
\keyword{plot}
\keyword{point}
\keyword{terrain}
\keyword{trees}
\keyword{vertical}
\keyword{voxel}
