% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.point.cloud_function.R
\name{thin.point.cloud}
\alias{thin.point.cloud}
\title{Point cloud thinning}
\usage{
thin.point.cloud(pc, cell.size = 1, returns.per.cell = 4)
}
\arguments{
\item{pc}{Point cloud in XYZ-table format (important: column names need to be X, Y and Z)}

\item{cell.size}{Side length of a spatial unit}

\item{returns.per.cell}{Number of returns that should be kept in each cell}
}
\value{
data.frame of the thinned point cloud in XYZ-table format
}
\description{
Function that thins a point cloud to a desired homogenous
horizontal point density by keeping a random subset of points
of a desired size in each spatial cell and discarding all other points.
}
\examples{
in progress
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{cloud}
\keyword{density}
\keyword{point}
\keyword{random}
\keyword{subsample}
\keyword{subset}
\keyword{thinning}

