% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_point_cloud_xyz_function.R
\name{thin_point_cloud_xyz}
\alias{thin_point_cloud_xyz}
\title{Point cloud thinning in 1D, 2D or 3D}
\usage{
thin_point_cloud_xyz(pc, res = 1, n = 4, dim = "xy")
}
\arguments{
\item{pc}{Point cloud in data.table, data.frame or LAS format (important: column names need to be X, Y and Z)}

\item{res}{Side length of a spatial unit (equal in all dimensions)}

\item{n}{Number of points to keep in each cell}

\item{dim}{Axes that define the thinning units, e.g. "xyz" for 3D voxels, "xy" for 2D grid cells in the horizontal plane,...}
}
\value{
Thinned point cloud in same format as input
}
\description{
Function that thins a point cloud to a desired homogenous
point density by keeping a random subset of n points
in each spatial unit and discarding all other points.
Spatial units are defined by their dimensions, i.e. 3D voxels (xyz),
2D cells (most commonly in xy plane), or can even be just along one
dimension (e.g. slices in z direction).
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
