% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical.foliage.profile_function.R
\name{vertical.foliage.profile}
\alias{vertical.foliage.profile}
\title{Derive a vertical foliage profile (VFP) from a Lidar profile}
\usage{
vertical.foliage.profile(profile, h.bin = 1, GR.threshold = 5, k = 0.5)
}
\arguments{
\item{profile}{Vector of Lidar profile (with heights as names)}

\item{h.bin}{Vertical extent of one height bin}

\item{GR.threshold}{Height threshold below which everything is regarded as ground
return (GR) and ignored in the processing}

\item{k}{Light extinction coefficient, which is an adjustment factor from relative
to absolute (LAD) VFP}
}
\value{
Vector of the vertical foliage profile (VFP)
}
\description{
Function that returns the vertical foliage profile for a given Lidar profile
using the approach of Harding et al. (2001) which is based on the MacArthur-Horn
method.
}
\examples{
prof.vec <- c(9, 2, 4, 5, 6, 3, 1)
names(prof.vec) <- 0:6
vfp.vec <- vertical.foliage.profile(profile=prof.vec, h.bin=1, GR.threshold=1, k=0.5)
plot(names(prof.vec) ~ prof.vec, type="l", xlab="Lidar returns", ylab="Height [m]")
plot(names(vfp.vec) ~ vfp.vec, type="l", xlab="LAD [m2/m3]", ylab="Height [m]")
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{Horn}
\keyword{LAD}
\keyword{LAI}
\keyword{MacArthur}
\keyword{adjustment}
\keyword{area}
\keyword{canopy}
\keyword{correction}
\keyword{cover}
\keyword{foliage}
\keyword{leaf}
\keyword{occlusion}
\keyword{profile}
\keyword{vegetation}

