% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel.array.from.point.cloud_function.R
\name{voxel.array.from.point.cloud}
\alias{voxel.array.from.point.cloud}
\title{Convert a point cloud to a voxel array}
\usage{
voxel.array.from.point.cloud(XYZ.df, global.h.max = NA, res.xy = 5,
  res.z = 5, value.var = "Z", func = "count")
}
\arguments{
\item{XYZ.df}{Lidar point cloud in XYZ-data.frame format}

\item{global.h.max}{Maximal Z of the array (if not set, the maximal Z of the point cloud is taken)}

\item{res.xy}{Horizontal sidelength of the voxels}

\item{res.z}{Vertical sidelength of the voxels}

\item{value.var}{Variable (data.frame column) from which to derive the voxel values}

\item{fun}{Aggregation function ("bool", "count", "min", "mean", "max", "sum")}
}
\value{
3D array
}
\description{
Function that takes a point cloud in XYZ-data.frame format and returns
a 3D array of custom horizontal and vertical voxel resolution.
}
\examples{
arr <- voxel.array.from.point.cloud(XYZ.df=pc, res.xy=10, res.z=1)
}
\keyword{Lidar}
\keyword{XYZ}
\keyword{array}
\keyword{canopy}
\keyword{cloud}
\keyword{empty}
\keyword{forest}
\keyword{point}
\keyword{space}
\keyword{structure}
\keyword{volume}
\keyword{voxel}
