% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_returns_point_cloud_function.R
\name{weighted_returns_point_cloud}
\alias{weighted_returns_point_cloud}
\title{Assign a weight to each point in a point cloud depending on total count points in voxel column}
\usage{
weighted_returns_point_cloud(pc, cell.size = 1)
}
\arguments{
\item{pc}{Data frame of the point cloud with first 3 columns representing X-, Y- and Z-coordinates}

\item{cell.size}{Horizontal side length of one spatial subunit (voxel column width in X- and Y-direction)}
}
\value{
Data frame of the point cloud with additional column "ReturnWeight"
}
\description{
Function that normalizes return weights in point clouds by dividing
the value 1 by the total count of returns in a voxel column of a given
horizontal resolution, e.g. if there are 4 returns in one vertical column
each gets a weight of 0.25
}
\examples{
in progress
}
\author{
Nikolai Knapp
}
\keyword{cloud}
\keyword{correction}
\keyword{density}
\keyword{lidar}
\keyword{normalization}
\keyword{point}
\keyword{profile}
\keyword{weighting}
\keyword{xyz}
